DockerAndDeployment/pyproject.toml
[project]
name = "restaurant-menu-system"
version = "3.0.0"
dependencies = [
    "fastapi>=0.104.0",
    "celery[redis]>=5.3.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "alembic>=1.12.0",
    "uvicorn[standard]>=0.23.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.25.0",
    # ... other dependencies as needed
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]